refactor(team): split lifecycle methods into launcher_lifecycle.go (C12)
Mechanical file split. The seven Launcher lifecycle methods that lived
in launcher.go (Launch, Attach, Kill, ResetSession, ReconfigureSession,
reconfigureVisibleAgents, respawnPanesAfterReseed) move into
launcher_lifecycle.go so the orchestrator file stays focused on
construction (NewLauncher) and shared state (Launcher struct, lazy
accessors).

Methods moved:

  Launch                     -> 150 lines, broker+tmux+goroutine boot
  Attach                     -> tmux -CC vs plain attach (iTerm aware)
  Kill                       -> drain workers, kill tmux session
  ResetSession               -> POST /reset to broker + clear state
  ReconfigureSession         -> reload config, rebuild manifest
  reconfigureVisibleAgents   -> respawn agent panes against new roster
  respawnPanesAfterReseed    -> office_reseeded notification handler

launcher.go: 1883 -> 1523 lines (-360). Cumulative since main:
4998 -> 1523 = -3475 lines (-69.5%).

Imports trimmed: launcher.go drops "log" + internal/provider (only
used by lifecycle methods). launcher_lifecycle.go imports its own
minimal set.

No new tests — pure file split. Existing test suite (32 packages
green) is the safety net. Each function appears in exactly one file
before, exactly one file after, with the same body.

Stacked on refactor/launcher-c11-notifier-split (C11).
